+++ Sources/VComponents/Components/Indicators/Page Indicator/Automatic/VPageIndicatorAutomaticUIModel.swift
@@ -6,6 +6,7 @@
 //
 
 import SwiftUI
+import VCore
 
 // MARK: - V Page Indicator Automatic UI Model
 /// Model that describes UI.
@@ -31,8 +32,8 @@ public struct VPageIndicatorAutomaticUIModel {
     /// Sub-model containing layout properties.
     public struct Layout {
         // MARK: Properties
-        /// Axis. Defaults to `horizontal`.
-        public var axis: Axis = pageIndicatorFiniteReference.layout.axis
+        /// Direction. Defaults to `leftToRight`.
+        public var direction: OmniLayoutDirection = pageIndicatorFiniteReference.layout.direction
         
         /// Number of visible dots. Default to `7`.
         ///
@@ -78,7 +79,7 @@ public struct VPageIndicatorAutomaticUIModel {
     var finiteSubModel: VPageIndicatorFiniteUIModel {
         var uiModel: VPageIndicatorFiniteUIModel = .init()
         
-        uiModel.layout.axis = layout.axis
+        uiModel.layout.direction = layout.direction
         uiModel.layout.dotDimension = layout.dotDimension
         uiModel.layout.unselectedDotScale = layout.unselectedDotScaleForFiniteConfiguration
         uiModel.layout.spacing = layout.spacing
@@ -93,7 +94,7 @@ public struct VPageIndicatorAutomaticUIModel {
     var infiniteSubModel: VPageIndicatorInfiniteUIModel {
         var uiModel: VPageIndicatorInfiniteUIModel = .init()
         
-        uiModel.layout.axis = layout.axis
+        uiModel.layout.direction = layout.direction
         uiModel.layout.visibleDots = layout.visibleDots
         uiModel.layout.centerDots = layout.centerDots
         uiModel.layout.dotDimension = layout.dotDimension
@@ -102,27 +103,6 @@ public struct VPageIndicatorAutomaticUIModel {
         uiModel.colors = colors
         
         uiModel.animations = animations
-        
-        return uiModel
-    }
-}
-
-// MARK: - Factory
-extension VPageIndicatorAutomaticUIModel {
-    /// `VPageIndicatorFiniteUIModel` with horizontal layout.
-    public static var horizontal: VPageIndicatorAutomaticUIModel {
-        var uiModel: VPageIndicatorAutomaticUIModel = .init()
-        
-        uiModel.layout.axis = .horizontal
-        
-        return uiModel
-    }
-    
-    /// `VPageIndicatorFiniteUIModel` with vertical layout.
-    public static var vertical: VPageIndicatorAutomaticUIModel {
-        var uiModel: VPageIndicatorAutomaticUIModel = .init()
-        
-        uiModel.layout.axis = .vertical
         
         return uiModel
     }